+++ nengo_loihi/config.py
@@ -42,4 +42,4 @@ def set_defaults():
 
     """
     nengo.Ensemble.max_rates.default = nengo.dists.Uniform(100, 120)
-    nengo.Ensemble.intercepts.default = nengo.dists.Uniform(-0.5, 0.5)
+    nengo.Ensemble.intercepts.default = nengo.dists.Uniform(-1.0, 0.5)